BDOG-2686 Don't update deployedConfig when the current deployedConfig lastUpdated field is more recent than the dataTimestamp
It was observed that if a deployment event is triggered after the scheduler has begun. The scheduler (which now has stale data
relative to the new deployment event), will overwrite the deployedConfig update triggered by the deployment event. This is because
the configIDs were different.

We now capture the timestamp when a deployment event is triggered, or when the scheduler begins, and pass this to our updateDeployment
method. Even if the configIDs differ, it will only update the deployedConfig if the current deployedConfig lastUpdated field is
prior to the timestamp.